Renamed Subject to Stream
Just to get the feeling if it’s a good idea or not
