+++ tests/test_streams/test_behaviorstream.py
@@ -1,8 +1,7 @@
-import sys
 from nose.tools import raises
 
 from rx.testing import TestScheduler, ReactiveTest
-from rx.subjects import ReplaySubject
+from rx.streams import BehaviorStream
 from rx.internal.exceptions import DisposedException
 
 on_next = ReactiveTest.on_next
@@ -41,7 +40,7 @@ def test_infinite():
         on_next(1020, 12)
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -52,11 +51,11 @@ def test_infinite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        stream[0] = BehaviorStream(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -64,15 +63,15 @@ def test_infinite():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -94,118 +93,24 @@ def test_infinite():
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(301, 3),
-        on_next(302, 4),
-        on_next(341, 5),
-        on_next(411, 6),
-        on_next(521, 7)
-    )
-
-    results2.messages.assert_equal(
-        on_next(401, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_next(631, 8)
-    )
-
-    results3.messages.assert_equal(
-        on_next(901, 10),
-        on_next(941, 11)
-    )
-
-
-def test_infinite2():
-    scheduler = TestScheduler()
-
-    xs = scheduler.create_hot_observable(
-        on_next(70, 1),
-        on_next(110, 2),
-        on_next(220, 3),
-        on_next(270, 4),
-        on_next(280, -1),
-        on_next(290, -2),
+        on_next(300, 4),
         on_next(340, 5),
         on_next(410, 6),
-        on_next(520, 7),
-        on_next(630, 8),
-        on_next(710, 9),
-        on_next(870, 10),
-        on_next(940, 11),
-        on_next(1020, 12)
-    )
-
-    subject = [None]
-    subscription = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
-
-    results1 = scheduler.create_observer()
-    results2 = scheduler.create_observer()
-    results3 = scheduler.create_observer()
-
-    def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
-    scheduler.schedule_absolute(100, action1)
-
-    def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
-    def action3(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action3)
-
-    def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action4)
-
-    def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action5)
-
-    def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action6)
-
-    def action7(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action7)
-
-    def action8(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action8)
-
-    def action9(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action9)
-
-    def action10(scheduler, state=None):
-        subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action10)
-
-    scheduler.start()
-
-    results1.messages.assert_equal(
-        on_next(301, 4),
-        on_next(302, -1),
-        on_next(303, -2),
-        on_next(341, 5),
-        on_next(411, 6),
-        on_next(521, 7)
+        on_next(520, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(401, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_next(631, 8)
+        on_next(400, 5),
+        on_next(410, 6),
+        on_next(520, 7),
+        on_next(630, 8)
     )
 
     results3.messages.assert_equal(
-        on_next(901, 10),
-        on_next(941, 11)
+        on_next(900, 10),
+        on_next(940, 11)
     )
+
 
 def test_finite():
     scheduler = TestScheduler()
@@ -221,10 +126,10 @@ def test_finite():
         on_completed(630),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, 'ex')
+        on_error(660, RxException())
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -235,71 +140,70 @@ def test_finite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        stream[0] = BehaviorStream(100)
     scheduler.schedule_absolute(100, action1)
 
+    def action2(scheduler, state=None):
+        subscription[0] = xs.subscribe(stream[0])
+    scheduler.schedule_absolute(200, action2)
+
     def action3(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action3)
+        subscription[0].dispose()
+    scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action4)
+        subscription1[0] = stream[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action5)
+        subscription2[0] = stream[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action6)
+        subscription3[0] = stream[0].subscribe(results3)
+    scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action7)
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(600, action7)
 
     def action8(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action8)
+        subscription2[0].dispose()
+    scheduler.schedule_absolute(700, action8)
 
     def action9(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action9)
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(800, action9)
 
     def action10(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action10)
-
-    def action11(scheduler, state=None):
         subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action11)
+    scheduler.schedule_absolute(950, action10)
 
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(301, 3),
-        on_next(302, 4),
-        on_next(341, 5),
-        on_next(411, 6),
-        on_next(521, 7)
+        on_next(300, 4),
+        on_next(340, 5),
+        on_next(410, 6),
+        on_next(520, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(401, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_completed(631)
+        on_next(400, 5),
+        on_next(410, 6),
+        on_next(520, 7),
+        on_completed(630)
     )
 
     results3.messages.assert_equal(
-        on_completed(901)
+        on_completed(900)
     )
 
 
 def test_error():
     scheduler = TestScheduler()
 
-    ex = RxException('ex')
+    ex = RxException()
 
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
@@ -312,10 +216,10 @@ def test_error():
         on_error(630, ex),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException('ex'))
+        on_error(660, RxException())
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -326,11 +230,11 @@ def test_error():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        stream[0] = BehaviorStream(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -338,15 +242,15 @@ def test_error():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -368,24 +272,22 @@ def test_error():
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(301, 3),
-        on_next(302, 4),
-        on_next(341, 5),
-        on_next(411, 6),
-        on_next(521, 7)
+        on_next(300, 4),
+        on_next(340, 5),
+        on_next(410, 6),
+        on_next(520, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(401, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_error(631, ex)
+        on_next(400, 5),
+        on_next(410, 6),
+        on_next(520, 7),
+        on_error(630, ex)
     )
 
     results3.messages.assert_equal(
-        on_error(901, ex)
+        on_error(900, ex)
     )
-
 
 def test_canceled():
     scheduler = TestScheduler()
@@ -397,7 +299,7 @@ def test_canceled():
         on_error(660, RxException())
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -408,11 +310,11 @@ def test_canceled():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        stream[0] = BehaviorStream(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -420,15 +322,15 @@ def test_canceled():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -450,42 +352,46 @@ def test_canceled():
     scheduler.start()
 
     results1.messages.assert_equal(
+        on_next(300, 100)
     )
 
     results2.messages.assert_equal(
-        on_completed(631)
+        on_next(400, 100),
+        on_completed(630)
     )
 
     results3.messages.assert_equal(
-        on_completed(901)
+        on_completed(900)
     )
 
-def test_subject_disposed():
+def test_stream_disposed():
     scheduler = TestScheduler()
 
-    subject = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
+    stream = [None]
 
     results1 = scheduler.create_observer()
+    subscription1 = [None]
+
     results2 = scheduler.create_observer()
+    subscription2 = [None]
+
     results3 = scheduler.create_observer()
+    subscription3 = [None]
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(scheduler=scheduler)
+        stream[0] = BehaviorStream(0)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(300, action3)
 
     def action4(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(400, action4)
 
     def action5(scheduler, state=None):
@@ -493,7 +399,7 @@ def test_subject_disposed():
     scheduler.schedule_absolute(500, action5)
 
     def action6(scheduler, state=None):
-        subject[0].dispose()
+        stream[0].dispose()
     scheduler.schedule_absolute(600, action6)
 
     def action7(scheduler, state=None):
@@ -505,142 +411,63 @@ def test_subject_disposed():
     scheduler.schedule_absolute(800, action8)
 
     def action9(scheduler, state=None):
-        subject[0].on_next(1)
+        stream[0].on_next(1)
     scheduler.schedule_absolute(150, action9)
 
     def action10(scheduler, state=None):
-        subject[0].on_next(2)
+        stream[0].on_next(2)
     scheduler.schedule_absolute(250, action10)
 
     def action11(scheduler, state=None):
-        subject[0].on_next(3)
+        stream[0].on_next(3)
     scheduler.schedule_absolute(350, action11)
 
     def action12(scheduler, state=None):
-        subject[0].on_next(4)
+        stream[0].on_next(4)
     scheduler.schedule_absolute(450, action12)
 
     def action13(scheduler, state=None):
-        subject[0].on_next(5)
+        stream[0].on_next(5)
     scheduler.schedule_absolute(550, action13)
 
     @raises(DisposedException)
     def action14(scheduler, state=None):
-        subject[0].on_next(6)
+        stream[0].on_next(6)
     scheduler.schedule_absolute(650, action14)
 
     @raises(DisposedException)
     def action15(scheduler, state=None):
-        subject[0].on_completed()
+        stream[0].on_completed()
     scheduler.schedule_absolute(750, action15)
 
     @raises(DisposedException)
     def action16(scheduler, state=None):
-        subject[0].on_error(Exception())
-
+        stream[0].on_error(RxException())
     scheduler.schedule_absolute(850, action16)
 
     @raises(DisposedException)
     def action17(scheduler, state=None):
-        subject[0].subscribe(None)
-
+        stream[0].subscribe(None)
     scheduler.schedule_absolute(950, action17)
 
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(201, 1),
-        on_next(251, 2),
-        on_next(351, 3),
-        on_next(451, 4)
+        on_next(200, 1),
+        on_next(250, 2),
+        on_next(350, 3),
+        on_next(450, 4)
     )
 
     results2.messages.assert_equal(
-        on_next(301, 1),
-        on_next(302, 2),
-        on_next(351, 3),
-        on_next(451, 4),
-        on_next(551, 5)
+        on_next(300, 2),
+        on_next(350, 3),
+        on_next(450, 4),
+        on_next(550, 5)
     )
 
     results3.messages.assert_equal(
-        on_next(401, 1),
-        on_next(402, 2),
-        on_next(403, 3),
-        on_next(451, 4),
-        on_next(551, 5)
+        on_next(400, 3),
+        on_next(450, 4),
+        on_next(550, 5)
     )
-
-
-def test_replay_subject_dies_out():
-    scheduler = TestScheduler()
-
-    xs = scheduler.create_hot_observable(
-        on_next(70, 1),
-        on_next(110, 2),
-        on_next(220, 3),
-        on_next(270, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(580)
-    )
-
-    subject = [None]
-
-    results1 = scheduler.create_observer()
-    results2 = scheduler.create_observer()
-    results3 = scheduler.create_observer()
-    results4 = scheduler.create_observer()
-
-    def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(sys.maxsize, 100, scheduler)
-    scheduler.schedule_absolute(100, action1)
-
-    def action2(scheduler, state=None):
-        xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
-    def action3(scheduler, state=None):
-        subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action3)
-
-    def action4(scheduler, state=None):
-        subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action4)
-
-    def action5(scheduler, state=None):
-        subject[0].subscribe(results3)
-    scheduler.schedule_absolute(600, action5)
-
-    def action6(scheduler, state=None):
-        subject[0].subscribe(results4)
-    scheduler.schedule_absolute(900, action6)
-
-    scheduler.start()
-
-    results1.messages.assert_equal(
-        on_next(301, 3),
-        on_next(302, 4),
-        on_next(341, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_completed(581)
-    )
-
-    results2.messages.assert_equal(
-        on_next(401, 5),
-        on_next(411, 6),
-        on_next(521, 7),
-        on_completed(581)
-    )
-
-    results3.messages.assert_equal(
-        on_next(601, 7),
-        on_completed(602)
-    )
-
-    results4.messages.assert_equal(
-        on_completed(901)
-    )
-
