+++ tests/test_streams/test_replaystream.py
@@ -1,7 +1,8 @@
+import sys
 from nose.tools import raises
 
 from rx.testing import TestScheduler, ReactiveTest
-from rx.subjects import BehaviorSubject
+from rx.streams import ReplayStream
 from rx.internal.exceptions import DisposedException
 
 on_next = ReactiveTest.on_next
@@ -40,7 +41,7 @@ def test_infinite():
         on_next(1020, 12)
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -51,11 +52,11 @@ def test_infinite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        stream[0] = ReplayStream(3, 100, scheduler)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -63,15 +64,15 @@ def test_infinite():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -93,24 +94,118 @@ def test_infinite():
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)
+        on_next(301, 3),
+        on_next(302, 4),
+        on_next(341, 5),
+        on_next(411, 6),
+        on_next(521, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_next(630, 8)
+        on_next(401, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_next(631, 8)
     )
 
     results3.messages.assert_equal(
-        on_next(900, 10),
-        on_next(940, 11)
+        on_next(901, 10),
+        on_next(941, 11)
     )
 
+
+def test_infinite2():
+    scheduler = TestScheduler()
+
+    xs = scheduler.create_hot_observable(
+        on_next(70, 1),
+        on_next(110, 2),
+        on_next(220, 3),
+        on_next(270, 4),
+        on_next(280, -1),
+        on_next(290, -2),
+        on_next(340, 5),
+        on_next(410, 6),
+        on_next(520, 7),
+        on_next(630, 8),
+        on_next(710, 9),
+        on_next(870, 10),
+        on_next(940, 11),
+        on_next(1020, 12)
+    )
+
+    stream = [None]
+    subscription = [None]
+    subscription1 = [None]
+    subscription2 = [None]
+    subscription3 = [None]
+
+    results1 = scheduler.create_observer()
+    results2 = scheduler.create_observer()
+    results3 = scheduler.create_observer()
+
+    def action1(scheduler, state=None):
+        stream[0] = ReplayStream(3, 100, scheduler)
+    scheduler.schedule_absolute(100, action1)
+
+    def action2(scheduler, state=None):
+        subscription[0] = xs.subscribe(stream[0])
+    scheduler.schedule_absolute(200, action2)
+
+    def action3(scheduler, state=None):
+        subscription[0].dispose()
+    scheduler.schedule_absolute(1000, action3)
+
+    def action4(scheduler, state=None):
+        subscription1[0] = stream[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action4)
+
+    def action5(scheduler, state=None):
+        subscription2[0] = stream[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action5)
+
+    def action6(scheduler, state=None):
+        subscription3[0] = stream[0].subscribe(results3)
+    scheduler.schedule_absolute(900, action6)
+
+    def action7(scheduler, state=None):
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(600, action7)
+
+    def action8(scheduler, state=None):
+        subscription2[0].dispose()
+    scheduler.schedule_absolute(700, action8)
+
+    def action9(scheduler, state=None):
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(800, action9)
+
+    def action10(scheduler, state=None):
+        subscription3[0].dispose()
+    scheduler.schedule_absolute(950, action10)
+
+    scheduler.start()
+
+    results1.messages.assert_equal(
+        on_next(301, 4),
+        on_next(302, -1),
+        on_next(303, -2),
+        on_next(341, 5),
+        on_next(411, 6),
+        on_next(521, 7)
+    )
+
+    results2.messages.assert_equal(
+        on_next(401, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_next(631, 8)
+    )
+
+    results3.messages.assert_equal(
+        on_next(901, 10),
+        on_next(941, 11)
+    )
 
 def test_finite():
     scheduler = TestScheduler()
@@ -126,10 +221,10 @@ def test_finite():
         on_completed(630),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException())
+        on_error(660, 'ex')
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -140,70 +235,71 @@ def test_finite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        stream[0] = ReplayStream(3, 100, scheduler)
     scheduler.schedule_absolute(100, action1)
 
-    def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
     def action3(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action3)
+        subscription[0] = xs.subscribe(stream[0])
+    scheduler.schedule_absolute(200, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action4)
+        subscription[0].dispose()
+    scheduler.schedule_absolute(1000, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action5)
+        subscription1[0] = stream[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action6)
+        subscription2[0] = stream[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action6)
 
     def action7(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action7)
+        subscription3[0] = stream[0].subscribe(results3)
+    scheduler.schedule_absolute(900, action7)
 
     def action8(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action8)
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(600, action8)
 
     def action9(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action9)
+        subscription2[0].dispose()
+    scheduler.schedule_absolute(700, action9)
 
     def action10(scheduler, state=None):
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(800, action10)
+
+    def action11(scheduler, state=None):
         subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action10)
+    scheduler.schedule_absolute(950, action11)
 
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)
+        on_next(301, 3),
+        on_next(302, 4),
+        on_next(341, 5),
+        on_next(411, 6),
+        on_next(521, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(630)
+        on_next(401, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_completed(631)
     )
 
     results3.messages.assert_equal(
-        on_completed(900)
+        on_completed(901)
     )
 
 
 def test_error():
     scheduler = TestScheduler()
 
-    ex = RxException()
+    ex = RxException('ex')
 
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
@@ -216,10 +312,10 @@ def test_error():
         on_error(630, ex),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException())
+        on_error(660, RxException('ex'))
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -230,11 +326,11 @@ def test_error():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        stream[0] = ReplayStream(3, 100, scheduler)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -242,15 +338,15 @@ def test_error():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -272,22 +368,24 @@ def test_error():
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)
+        on_next(301, 3),
+        on_next(302, 4),
+        on_next(341, 5),
+        on_next(411, 6),
+        on_next(521, 7)
     )
 
     results2.messages.assert_equal(
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_error(630, ex)
+        on_next(401, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_error(631, ex)
     )
 
     results3.messages.assert_equal(
-        on_error(900, ex)
+        on_error(901, ex)
     )
+
 
 def test_canceled():
     scheduler = TestScheduler()
@@ -299,7 +397,7 @@ def test_canceled():
         on_error(660, RxException())
     )
 
-    subject = [None]
+    stream = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
@@ -310,11 +408,11 @@ def test_canceled():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        stream[0] = ReplayStream(3, 100, scheduler)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
+        subscription[0] = xs.subscribe(stream[0])
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
@@ -322,15 +420,15 @@ def test_canceled():
     scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
@@ -352,46 +450,42 @@ def test_canceled():
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(300, 100)
     )
 
     results2.messages.assert_equal(
-        on_next(400, 100),
-        on_completed(630)
+        on_completed(631)
     )
 
     results3.messages.assert_equal(
-        on_completed(900)
+        on_completed(901)
     )
 
-def test_subject_disposed():
+def test_stream_disposed():
     scheduler = TestScheduler()
 
-    subject = [None]
-
-    results1 = scheduler.create_observer()
+    stream = [None]
     subscription1 = [None]
-
-    results2 = scheduler.create_observer()
     subscription2 = [None]
-
-    results3 = scheduler.create_observer()
     subscription3 = [None]
 
+    results1 = scheduler.create_observer()
+    results2 = scheduler.create_observer()
+    results3 = scheduler.create_observer()
+
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(0)
+        stream[0] = ReplayStream(scheduler=scheduler)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
+        subscription1[0] = stream[0].subscribe(results1)
     scheduler.schedule_absolute(200, action2)
 
     def action3(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
+        subscription2[0] = stream[0].subscribe(results2)
     scheduler.schedule_absolute(300, action3)
 
     def action4(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
+        subscription3[0] = stream[0].subscribe(results3)
     scheduler.schedule_absolute(400, action4)
 
     def action5(scheduler, state=None):
@@ -399,7 +493,7 @@ def test_subject_disposed():
     scheduler.schedule_absolute(500, action5)
 
     def action6(scheduler, state=None):
-        subject[0].dispose()
+        stream[0].dispose()
     scheduler.schedule_absolute(600, action6)
 
     def action7(scheduler, state=None):
@@ -411,63 +505,142 @@ def test_subject_disposed():
     scheduler.schedule_absolute(800, action8)
 
     def action9(scheduler, state=None):
-        subject[0].on_next(1)
+        stream[0].on_next(1)
     scheduler.schedule_absolute(150, action9)
 
     def action10(scheduler, state=None):
-        subject[0].on_next(2)
+        stream[0].on_next(2)
     scheduler.schedule_absolute(250, action10)
 
     def action11(scheduler, state=None):
-        subject[0].on_next(3)
+        stream[0].on_next(3)
     scheduler.schedule_absolute(350, action11)
 
     def action12(scheduler, state=None):
-        subject[0].on_next(4)
+        stream[0].on_next(4)
     scheduler.schedule_absolute(450, action12)
 
     def action13(scheduler, state=None):
-        subject[0].on_next(5)
+        stream[0].on_next(5)
     scheduler.schedule_absolute(550, action13)
 
     @raises(DisposedException)
     def action14(scheduler, state=None):
-        subject[0].on_next(6)
+        stream[0].on_next(6)
     scheduler.schedule_absolute(650, action14)
 
     @raises(DisposedException)
     def action15(scheduler, state=None):
-        subject[0].on_completed()
+        stream[0].on_completed()
     scheduler.schedule_absolute(750, action15)
 
     @raises(DisposedException)
     def action16(scheduler, state=None):
-        subject[0].on_error(RxException())
+        stream[0].on_error(Exception())
+
     scheduler.schedule_absolute(850, action16)
 
     @raises(DisposedException)
     def action17(scheduler, state=None):
-        subject[0].subscribe(None)
+        stream[0].subscribe(None)
+
     scheduler.schedule_absolute(950, action17)
 
     scheduler.start()
 
     results1.messages.assert_equal(
-        on_next(200, 1),
-        on_next(250, 2),
-        on_next(350, 3),
-        on_next(450, 4)
+        on_next(201, 1),
+        on_next(251, 2),
+        on_next(351, 3),
+        on_next(451, 4)
     )
 
     results2.messages.assert_equal(
-        on_next(300, 2),
-        on_next(350, 3),
-        on_next(450, 4),
-        on_next(550, 5)
+        on_next(301, 1),
+        on_next(302, 2),
+        on_next(351, 3),
+        on_next(451, 4),
+        on_next(551, 5)
     )
 
     results3.messages.assert_equal(
-        on_next(400, 3),
-        on_next(450, 4),
-        on_next(550, 5)
+        on_next(401, 1),
+        on_next(402, 2),
+        on_next(403, 3),
+        on_next(451, 4),
+        on_next(551, 5)
     )
+
+
+def test_replay_stream_dies_out():
+    scheduler = TestScheduler()
+
+    xs = scheduler.create_hot_observable(
+        on_next(70, 1),
+        on_next(110, 2),
+        on_next(220, 3),
+        on_next(270, 4),
+        on_next(340, 5),
+        on_next(410, 6),
+        on_next(520, 7),
+        on_completed(580)
+    )
+
+    stream = [None]
+
+    results1 = scheduler.create_observer()
+    results2 = scheduler.create_observer()
+    results3 = scheduler.create_observer()
+    results4 = scheduler.create_observer()
+
+    def action1(scheduler, state=None):
+        stream[0] = ReplayStream(sys.maxsize, 100, scheduler)
+    scheduler.schedule_absolute(100, action1)
+
+    def action2(scheduler, state=None):
+        xs.subscribe(stream[0])
+    scheduler.schedule_absolute(200, action2)
+
+    def action3(scheduler, state=None):
+        stream[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action3)
+
+    def action4(scheduler, state=None):
+        stream[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action4)
+
+    def action5(scheduler, state=None):
+        stream[0].subscribe(results3)
+    scheduler.schedule_absolute(600, action5)
+
+    def action6(scheduler, state=None):
+        stream[0].subscribe(results4)
+    scheduler.schedule_absolute(900, action6)
+
+    scheduler.start()
+
+    results1.messages.assert_equal(
+        on_next(301, 3),
+        on_next(302, 4),
+        on_next(341, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_completed(581)
+    )
+
+    results2.messages.assert_equal(
+        on_next(401, 5),
+        on_next(411, 6),
+        on_next(521, 7),
+        on_completed(581)
+    )
+
+    results3.messages.assert_equal(
+        on_next(601, 7),
+        on_completed(602)
+    )
+
+    results4.messages.assert_equal(
+        on_completed(901)
+    )
+
